Changes to add application status

+++ src/main/java/com/fyp/companyinterfaceservice/model/Stats.java
@@ -10,20 +10,20 @@ import java.util.Map;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Stats {
-    private int male;
-    private int female;
-    private int ageTier1;
-    private int ageTier2;
-    private int ageTier3;
-    private int ageTier4;
-    private int ageTier5;
-    private int ageTier6;
-    private int ageTier7;
-    private int white;
-    private int bAA;
-    private int aIAN;
-    private int asian;
-    private int nHPI;
-    private int mR;
-    private Map<String, Integer> courses;
+    private double male;
+    private double female;
+    private double ageTier1;
+    private double ageTier2;
+    private double ageTier3;
+    private double ageTier4;
+    private double ageTier5;
+    private double ageTier6;
+    private double ageTier7;
+    private double white;
+    private double bAA;
+    private double aIAN;
+    private double asian;
+    private double nHPI;
+    private double mR;
+    private Map<String, Double> courses;
 }
